Refactor duplicate code
